+++ systemimage/testing/dbus.py
@@ -77,6 +77,7 @@ class _LiveTestableService(Service):
         self._downloading = False
         self._rebootable = False
         self._failure_count = 0
+        del config.build_number
         safe_remove(config.system.state_file)
         safe_remove(config.system.settings_db)